fix : 지역 인증 레포지토리 db스펙과 다른 오류 픽스

+++ meerket/meerket-domain/src/main/java/org/j1p5/domain/areaAuth/repository/AreaAuthRepository.java
@@ -3,6 +3,6 @@ package org.j1p5.domain.areaAuth.repository;
 import org.j1p5.domain.areaAuth.entity.AreaAuthEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AreaAuthRepository extends JpaRepository<AreaAuthEntity, Long> {
+public interface AreaAuthRepository extends JpaRepository<AreaAuthEntity, Integer> {
     boolean existsByUserId(Long userId);
 }
